Trying to recoller les morceaux as we like to say

+++ app/src/main/java/com/example/projet_progm/ActivityMemento.kt
@@ -46,7 +46,7 @@ class ActivityMemento : ComponentActivity() {
     private val endGame = object : Runnable {
         override fun run() {
             val resultIntent = Intent()
-            resultIntent.putExtra("score", score)
+            resultIntent.putExtra("score", score.toInt())
             setResult(Activity.RESULT_OK, resultIntent)
             finish()
         }
