+++ app/src/main/java/com/example/projet_progm/QuestionnaireGameActivity.kt
@@ -163,7 +163,7 @@ class QuestionnaireGameActivity : ComponentActivity() {
     private val endGame = object : Runnable {
         override fun run() {
             val resultIntent = Intent()
-            resultIntent.putExtra("score", score)
+            resultIntent.putExtra("score", score.toInt())
             setResult(Activity.RESULT_OK, resultIntent)
             finish()
         }
